pyproject.toml
[tool.poetry]
name = "idiotDiaryV2"
version = "0.1.0"
description = ""
authors = ["rSlow <redmaxx4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "3.4.1"
aiogram-dialog = "2.2.0a3"
alembic = "1.13.1"
annotated-types = "0.6.0"
apscheduler = "3.10.4"
asyncpg = "0.29.0"
babel = "2.15.0"
environs = "11.0.0"
eyed3 = "0.9.7"
jinja2 = "3.1.4"
loguru = "0.7.2"
pillow = "10.3.0"
pytz = "2024.1"
redis = "5.1.0b4"
selenium = "4.19.0"
sqlalchemy = "2.0.29"
typing-extensions = "4.11.0"
yt-dlp = "2024.4.9"
fastapi = "^0.111.1"
adaptix = "^3.0.0b7"
asgi-monitor = "^0.4.0"
dishka = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
